requirements.txt
annotated-types==0.7.0
blinker==1.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.1.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
eval_type_backport==0.2.2
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.9
httplib2==0.20.2
huggingface-hub==0.34.4
idna==3.10
imageio==2.35.1
importlib-metadata==4.6.4
importlib_resources==6.4.5
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.4.2
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.7.5
more-itertools==8.10.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.0.3
patsy==1.0.1
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.10.6
pydantic_core==2.27.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==3.3.1
sentry-sdk==2.32.0
setproctitle==1.3.6
six==1.16.0
smmap==5.0.2
statsmodels==0.14.1
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.5.0
tifffile==2023.7.10
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.168
ultralytics-thop==2.0.14
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
wandb==0.20.1
zipp==3.20.2
